Refactoring to align with the worker ecs flow
