+++ backend/dataall/modules/shares_base/api/resolvers.py
@@ -109,10 +109,6 @@ def reapply_items_share_object(context: Context, source, input):
     return ShareItemService.reapply_items_share_object(uri=share_uri, item_uris=reapply_item_uris)
 
 
-def reapply_share_items_share_object_for_dataset(context: Context, source, datasetUri):
-    return ShareItemService.reapply_share_items_for_dataset(uri=datasetUri)
-
-
 def delete_share_object(context: Context, source, shareUri: str = None):
     return ShareObjectService.delete_share_object(uri=shareUri)
